debug the controller and model
check the integer for accountID and book ID; add filter to the input name, no {"'", "\"", "\\", ";", "=", ">", "<","*"}

+++ LibraryManagementSystemProject/src/controller/LMSController.java
@@ -19,6 +19,17 @@ public class LMSController {
         return modelController;
     }
 
+    public static boolean isInteger(String s) {
+        try {
+            Integer.parseInt(s);
+        } catch(NumberFormatException e) {
+            return false;
+        } catch(NullPointerException e) {
+            return false;
+        }
+        return true;
+    }
+
     public static void main(String[] args) {
         while(true) {
             mainView.welcomePage();
@@ -28,14 +39,28 @@ public class LMSController {
                 mainView.mainPageWelcome();
                 // compulsory set date
                 System.out.println("Please set today's date first");
-                System.out.print(">>> Please enter the Year: ");
-                String y = inputListener();
-                System.out.print(">>> Please enter the Month: ");
-                String m = inputListener();
-                System.out.print(">>> Please enter the Day: ");
-                String d = inputListener();
-                modelController.setDate(y, m, d);
-
+                while(true) {
+                    System.out.print(">>> Please enter the Year: ");
+                    String y = inputListener();
+                    if (!isInteger(y)) {
+                        System.out.println("Invalid input, please enter again");
+                        continue;
+                    }
+                    System.out.print(">>> Please enter the Month: ");
+                    String m = inputListener();
+                    if (!isInteger(m)) {
+                        System.out.println("Invalid input, please enter again");
+                        continue;
+                    }
+                    System.out.print(">>> Please enter the Day: ");
+                    String d = inputListener();
+                    if (!isInteger(d)) {
+                        System.out.println("Invalid input, please enter again");
+                        continue;
+                    }
+                    modelController.setDate(y, m, d);
+                    break;
+                }
                 while (true) {
                     mainView.mainPage();
                     String mainOption = inputListener();
@@ -44,6 +69,10 @@ public class LMSController {
                             System.out.print("Please enter book ID (- [Back] for Back):");
                             String inBookID = inputListener();
                             if (inBookID.equals("Back")) break; // back to previous page (assumption: there is no input value named "Back")
+                            if (!isInteger(inBookID)) {
+                                System.out.println("Book ID should be an integer, please try again!");
+                                continue;
+                            }
                             System.out.print("Please enter book ISBN (- [Back] for Back):");
                             String inBookISBN = inputListener();
                             if (inBookISBN.equals("Back")) break; // back to previous page (assumption: there is no input value named "Back")
@@ -73,6 +102,10 @@ public class LMSController {
                             mainView.inputUserPage();
                             String accountID = inputListener();
                             if (accountID.equals("Back")) break; // back to previous page (assumption: there is no input value named "Back")
+                            if (!isInteger(accountID)) {
+                                System.out.println("Account ID should be an integer, please try again!");
+                                continue;
+                            }
                             User user = new User(accountID, true, "Notification \n");
                             if (modelController.addRecord(user)) {
                                 mainView.successPage();
@@ -89,6 +122,10 @@ public class LMSController {
                             mainView.inputBookIDPage();
                             String bookID = inputListener();
                             if (bookID.equals("Back")) break; // back to previous page (assumption: there is no input value named "Back")
+                            if (!isInteger(bookID)) {
+                                System.out.println("Book ID should be an integer, please try again!");
+                                continue;
+                            }
                             if (modelController.deleteBookRecord(bookID)) {
                                 mainView.successPage();
                                 mainView.linePage();
@@ -104,6 +141,10 @@ public class LMSController {
                             mainView.inputUserPage();
                             String accountID = inputListener();
                             if (accountID.equals("Back")) break; // back to previous page (assumption: there is no input value named "Back")
+                            if (!isInteger(accountID)) {
+                                System.out.println("Account ID should be an integer, please try again!");
+                                continue;
+                            }
                             if (modelController.deleteUserRecord(accountID)) {
                                 mainView.successPage();
                                 mainView.linePage();
@@ -120,7 +161,12 @@ public class LMSController {
                             String userAccountID = inputListener();
                             if (userAccountID.equals("Back"))
                                 break; // back to previous page (assumption: there is no input value named "Back")
+                            if (!isInteger(userAccountID)) {
+                                System.out.println("Account ID should be an integer, please try again!");
+                                continue;
+                            }
                             if (modelController.searchUserOnAccountID(userAccountID).size() == 0) {
+                                System.out.println("No such user found!");
                                 continue;
                             }
                             User user = modelController.searchUserOnAccountID(userAccountID).get(0);
@@ -128,6 +174,10 @@ public class LMSController {
                                 mainView.processUserRentPage();
                                 String inBookID = inputListener();
                                 if (inBookID.equals("Back")) break; // back to previous page (assumption: there is no input value named "Back")
+                                if (!isInteger(inBookID)) {
+                                    System.out.println("Book ID should be an integer, please try again!");
+                                    continue;
+                                }
                                 if (modelController.rentBookFromUser(userAccountID, inBookID)) { //used rentBookFromUser return boolean
                                     mainView.successPage();
                                     mainView.linePage();
@@ -149,15 +199,25 @@ public class LMSController {
                             String userAccountID = inputListener();
                             if (userAccountID.equals("Back"))
                                 break; // back to previous page (assumption: there is no input value named "Back")
+                            if (!isInteger(userAccountID)) {
+                                System.out.println("Account ID should be an integer, please try again!");
+                                continue;
+                            }
                             if (modelController.searchUserOnAccountID(userAccountID).size() == 0) {
+                                System.out.println("No such user found!");
                                 continue;
                             }
                             User user = modelController.searchUserOnAccountID(userAccountID).get(0);
                             if (user.getAccountStatus()) {
                                 mainView.processUserRentPage();
                                 String inBookID = inputListener();
-                                if (inBookID.equals("Back"))
+                                if (inBookID.equals("Back")) {
                                     break; // back to previous page (assumption: there is no input value named "Back")
+                                }
+                                if (!isInteger(inBookID)) {
+                                    System.out.println("Book ID should be an integer, please try again!");
+                                    continue;
+                                }
                                 if (modelController.rentBookFromPlacedBook(userAccountID, inBookID)) { //used rentBookFromUser return boolean
                                     mainView.successPage();
                                     mainView.linePage();
@@ -176,9 +236,15 @@ public class LMSController {
                         while (true) {
                             mainView.inputUserPage();
                             String userAccountID = inputListener();
-                            if (userAccountID.equals("Back"))
+                            if (userAccountID.equals("Back")) {
                                 break; // back to previous page (assumption: there is no input value named "Back")
+                            }
+                            if (!isInteger(userAccountID)) {
+                                System.out.println("Account ID should be an integer, please try again!");
+                                continue;
+                            }
                             if (modelController.searchUserOnAccountID(userAccountID).size() == 0) {
+                                System.out.println("No such user found!");
                                 continue;
                             }
                             User user = modelController.searchUserOnAccountID(userAccountID).get(0);
@@ -203,9 +269,15 @@ public class LMSController {
                         while (true) {
                             mainView.inputUserPage();
                             String userAccountID = inputListener();
-                            if (userAccountID.equals("Back"))
+                            if (userAccountID.equals("Back")) {
                                 break; // back to previous page (assumption: there is no input value named "Back")
+                            }
+                            if (!isInteger(userAccountID)) {
+                                System.out.println("Account ID should be an integer, please try again!");
+                                continue;
+                            }
                             if (modelController.searchUserOnAccountID(userAccountID).size() == 0) {
+                                System.out.println("No such user found!");
                                 continue;
                             }
                             User user = modelController.searchUserOnAccountID(userAccountID).get(0);
@@ -232,17 +304,28 @@ public class LMSController {
                         while (true) {
                             mainView.inputUserPage();
                             String userAccountID = inputListener();
-                            if (userAccountID.equals("Back"))
+                            if (userAccountID.equals("Back")) {
                                 break; // back to previous page (assumption: there is no input value named "Back")
+                            }
+                            if (!isInteger(userAccountID)) {
+                                System.out.println("Account ID should be an integer, please try again!");
+                                continue;
+                            }
                             if (modelController.searchUserOnAccountID(userAccountID).size() == 0) {
+                                System.out.println("No such user found!");
                                 continue;
                             }
                             User user = modelController.searchUserOnAccountID(userAccountID).get(0);
                             if (user.getAccountStatus()) {
                                 mainView.processUserCancelPlacedPage();
                                 String inBookID = inputListener();
-                                if (inBookID.equals("Back"))
+                                if (inBookID.equals("Back")) {
                                     break; // back to previous page (assumption: there is no input value named "Back")
+                                }
+                                if (!isInteger(inBookID)) {
+                                    System.out.println("Book ID should be an integer, please try again!");
+                                    continue;
+                                }
                                 if (modelController.cancelPlacedBook(inBookID, userAccountID)) {
                                     mainView.successPage();
                                     mainView.linePage();
@@ -367,7 +450,8 @@ public class LMSController {
                         mainView.inputUserPage();
                         String inAccountID = inputListener();
                         try {
-                            if (modelController.searchUserOnAccountID(inAccountID).size() == 0) {
+                            if (isInteger(inAccountID) && modelController.searchUserOnAccountID(inAccountID).size() == 0) {
+                                System.out.println("No such user found!");
                                 mainView.emptyPage();
                             } else {
                                 mainView.showListPage();
